pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datakhanon"
version = "0.1.0"
description = "Biblioteca para mineração de dados e aprendizado de máquina com fluxo end-to-end."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }

authors = [
  { name = "Vinicius Santos", email = "vinicius.santos@ifsp.edu.br" },
]

keywords = [
  "data-mining",
  "machine-learning",
  "feature-engineering",
  "preprocessing",
  "eda",
  "model-selection",
  "datakhanon",
]

# ==========================
# Dependências principais
# ==========================
# Necessário para:
# - pré-processamento (pandas, numpy, scikit-learn, scipy, imbalanced-learn)
# - modelagem clássica
# - relatórios/EDA (matplotlib, seaborn)
# - serialização/configuração (joblib, PyYAML)
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "scipy>=1.10",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "joblib>=1.3",
  "PyYAML>=6.0",
  "imbalanced-learn>=0.12",  # oversampling/undersampling em pipelines
]

# ==========================
# Metadados adicionais
# ==========================
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
Homepage = "https://github.com/ViniciusKanh/datakhanon"
"Bug Tracker" = "https://github.com/ViniciusKanh/datakhanon/issues"
Documentation = "https://github.com/ViniciusKanh/datakhanon"  # ajuste quando tiver docs separados
Source = "https://github.com/ViniciusKanh/datakhanon"

# ==========================
# Descoberta de pacotes
# ==========================
[tool.setuptools.packages.find]
where = ["."]
include = ["datakhanon*"]

[tool.setuptools]
include-package-data = true

# ==========================
# Dependências opcionais
# ==========================
[project.optional-dependencies]

# Visualização extra (além de matplotlib/seaborn)
viz = [
  "plotly>=5.0",
  "bokeh>=3.0",
]

# Dashboards interativos
dashboard = [
  "streamlit>=1.30",
]

# Explicabilidade de modelos
explainer = [
  "shap>=0.44",
]

# Redução de dimensionalidade / embedding para visualização
embed = [
  "umap-learn>=0.5",
]

# Experimentos de ensemble avançados (sua outra lib)
ensemble = [
  "auto-ensemble-benchmark @ https://github.com/ViniciusKanh/auto-ensemble-benchmark/releases/download/v0.1.3/auto_ensemble_benchmark-0.1.3-py3-none-any.whl"
]

# Dependências de desenvolvimento (não vão para o usuário final)
dev = [
  "black>=24.0",
  "isort>=5.13",
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "mypy>=1.8",
  "ipython>=8.0",
  "twine>=6.0",
  "build>=1.0",
]

# ==========================
# Ferramentas internas
# ==========================
[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
